+++ tests/test_wsgi.py
@@ -263,6 +263,7 @@ def test_get_current_url_invalid_utf8():
     assert rv == "http://localhost/?foo=bar&baz=blah&meh=%CF"
 
 
+@pytest.mark.filterwarnings("ignore:'make_line_iter:DeprecationWarning")
 def test_multi_part_line_breaks():
     data = "abcdef\r\nghijkl\r\nmnopqrstuvwxyz\r\nABCDEFGHIJK"
     test_stream = io.StringIO(data)
@@ -279,6 +280,7 @@ def test_multi_part_line_breaks():
     ]
 
 
+@pytest.mark.filterwarnings("ignore:'make_line_iter:DeprecationWarning")
 def test_multi_part_line_breaks_bytes():
     data = b"abcdef\r\nghijkl\r\nmnopqrstuvwxyz\r\nABCDEFGHIJK"
     test_stream = io.BytesIO(data)
@@ -300,6 +302,7 @@ def test_multi_part_line_breaks_bytes():
     ]
 
 
+@pytest.mark.filterwarnings("ignore:'make_line_iter:DeprecationWarning")
 def test_multi_part_line_breaks_problematic():
     data = "abc\rdef\r\nghi"
     for _ in range(1, 10):
@@ -308,12 +311,14 @@ def test_multi_part_line_breaks_problematic():
         assert lines == ["abc\r", "def\r\n", "ghi"]
 
 
+@pytest.mark.filterwarnings("ignore:'make_line_iter:DeprecationWarning")
 def test_iter_functions_support_iterators():
     data = ["abcdef\r\nghi", "jkl\r\nmnopqrstuvwxyz\r", "\nABCDEFGHIJK"]
     lines = list(wsgi.make_line_iter(data))
     assert lines == ["abcdef\r\n", "ghijkl\r\n", "mnopqrstuvwxyz\r\n", "ABCDEFGHIJK"]
 
 
+@pytest.mark.filterwarnings("ignore:'make_chunk_iter:DeprecationWarning")
 def test_make_chunk_iter():
     data = ["abcdefXghi", "jklXmnopqrstuvwxyzX", "ABCDEFGHIJK"]
     rv = list(wsgi.make_chunk_iter(data, "X"))
@@ -325,6 +330,7 @@ def test_make_chunk_iter():
     assert rv == ["abcdef", "ghijkl", "mnopqrstuvwxyz", "ABCDEFGHIJK"]
 
 
+@pytest.mark.filterwarnings("ignore:'make_chunk_iter:DeprecationWarning")
 def test_make_chunk_iter_bytes():
     data = [b"abcdefXghi", b"jklXmnopqrstuvwxyzX", b"ABCDEFGHIJK"]
     rv = list(wsgi.make_chunk_iter(data, "X"))
@@ -357,6 +363,7 @@ def test_make_chunk_iter_bytes():
     ]
 
 
+@pytest.mark.filterwarnings("ignore:'make_line_iter:DeprecationWarning")
 def test_lines_longer_buffer_size():
     data = "1234567890\n1234567890\n"
     for bufsize in range(1, 15):
@@ -366,6 +373,7 @@ def test_lines_longer_buffer_size():
         assert lines == ["1234567890\n", "1234567890\n"]
 
 
+@pytest.mark.filterwarnings("ignore:'make_line_iter:DeprecationWarning")
 def test_lines_longer_buffer_size_cap():
     data = "1234567890\n1234567890\n"
     for bufsize in range(1, 15):