+++ CHANGES.rst
@@ -14,6 +14,7 @@ Unreleased
     filenames in downloads. :issue:`2598`
 -   Deprecate the ``safe_conversion`` parameter of ``iri_to_uri``. The ``Location``
     header is converted to IRI using the same process as everywhere else. :issue:`2609`
+-   Deprecate ``werkzeug.wsgi.make_line_iter`` and ``make_chunk_iter``. :pr:`2613`
 -   Use modern packaging metadata with ``pyproject.toml`` instead of ``setup.cfg``.
     :pr:`2574`
 -   ``Request.get_json()`` will raise a ``415 Unsupported Media Type`` error if the